# perf(crypto-block): use workers to deserialize transactions