--- a/packages/contracts/source/contracts/crypto/transactions.ts
+++ b/packages/contracts/source/contracts/crypto/transactions.ts
@@ -97,7 +97,7 @@ export interface TransactionSerializer {
 }
 
 export interface TransactionDeserializer {
-	deserialize(serialized: string | Buffer): Promise<Transaction>;
+	deserialize(serialized: Buffer): Promise<Transaction>;
 }
 
 export interface TransactionFactory {