--- a/packages/contracts/source/contracts/crypto/transactions.ts
+++ b/packages/contracts/source/contracts/crypto/transactions.ts
@@ -75,6 +75,14 @@ export interface SerializeOptions {
 	// TODO: consider passing pre-allocated buffer
 }
 
+export interface TransactionCryptoData {
+	readonly hash: string;
+	readonly publicKey: string;
+	readonly address: string;
+	readonly legacyAddress?: string;
+	readonly schemaError?: any;
+}
+
 export interface TransactionServiceProvider {
 	register(): Promise<void>;
 }
@@ -108,6 +116,8 @@ export interface TransactionFactory {
 	fromJson(json: TransactionJson): Promise<Transaction>;
 
 	fromData(data: TransactionData, strict?: boolean): Promise<Transaction>;
+
+	computeCryptoData(data: TransactionData): Promise<TransactionCryptoData>;
 }
 
 export type TransactionConstructor = any;