--- a/packages/crypto-messages/source/factory.test.ts
+++ b/packages/crypto-messages/source/factory.test.ts
@@ -45,6 +45,11 @@ describe<{
 						context.sandbox.app
 							.getTagged(Identifiers.Cryptography.Signature.Instance, "type", "consensus")!
 							[method](message, privateKey),
+					// @ts-ignore
+					transactionFactory: (method, message, privateKey) =>
+						context.sandbox.app
+							.get(Identifiers.Cryptography.Transaction.Factory)!
+							[method](message, privateKey),
 				};
 			},
 		};