--- a/packages/crypto-messages/source/proposal.test.ts
+++ b/packages/crypto-messages/source/proposal.test.ts
@@ -30,6 +30,9 @@ describe<{
 					context.sandbox.app
 						.getTagged(Identifiers.Cryptography.Signature.Instance, "type", "consensus")!
 						[method](message, privateKey),
+				// @ts-ignore
+				transactionFactory: (method, message, privateKey) =>
+					context.sandbox.app.get(Identifiers.Cryptography.Transaction.Factory)![method](message, privateKey),
 			}),
 		};